+++ rules/filenames.js
@@ -1,0 +1,10 @@
+module.exports = {
+  plugins: [
+    "filenames"
+  ],
+  rules: {
+    "filenames/match-exported": "error",
+    "filenames/match-regex": "off",
+    "filenames/no-index": "error"
+  }
+};