+++ rules/react-hooks.js
@@ -1,0 +1,9 @@
+module.exports = {
+  plugins: [
+    "react-hooks"
+  ],
+  rules: {
+    "react-hooks/exhaustive-deps": "warn",
+    "react-hooks/rules-of-hooks": "error"
+  }
+};