+++ rules/variables.js
@@ -1,0 +1,26 @@
+const confusingBrowserGlobals = require("confusing-browser-globals");
+
+module.exports = {
+  rules: {
+    "init-declarations": "off",
+    "no-catch-shadow": "off",
+    "no-delete-var": "error",
+    "no-label-var": "error",
+    "no-restricted-globals": ["error", "isFinite", "isNaN"].concat(confusingBrowserGlobals),
+    "no-shadow": "error",
+    "no-shadow-restricted-names": "error",
+    "no-undef": "error",
+    "no-undef-init": "error",
+    "no-undefined": "off",
+    "no-unused-vars": ["error", {
+      argsIgnorePattern: "^_",
+      ignoreRestSiblings: true,
+      varsIgnorePattern: "^_"
+    }],
+    "no-use-before-define": ["error", {
+      functions: true,
+      classes: true,
+      variables: true
+    }]
+  }
+};